+++ NEWS.md
@@ -1,4 +1,4 @@
-# random.cdisc.data 0.3.10.9000
+# random.cdisc.data 0.3.11
 
 ### Enhancements
 * Added `AECONTRT` variable to ADAE dataset.